Replace variable types from g* to standard types.
replacing
- gchar - char
- gint - int, unsigned int, or size_t
- guint -> unsigned int or size_t
- guint8 -> unsigned char
- gboolean -> bool
- gsize -> size_t
- gpointer -> void *
- gconstpointer -> const void *
- gunichar -> unichar (typedef of int on PyZyUtil.h)

BUG=None
TEST=Run a test

Review URL: http://codereview.appspot.com/5367046

+++ src/PyZyFullPinyinContext.cc
@@ -97,7 +97,7 @@ FullPinyinContext::removeWordBefore (void)
     if (G_UNLIKELY (m_cursor == 0))
         return false;
 
-    guint cursor;
+    size_t cursor;
 
     if (G_UNLIKELY (m_cursor > m_pinyin_len)) {
         cursor = m_pinyin_len;
