+++ src/PyZyInputContext.h
@@ -78,10 +78,10 @@ public:
     virtual bool moveCursorToBegin (void) = 0;
     virtual bool moveCursorToEnd (void) = 0;
 
-    virtual bool selectCandidate (unsigned int index) = 0;
+    virtual bool selectCandidate (size_t index) = 0;
     virtual bool focusCandidatePrevious (void) = 0;
     virtual bool focusCandidateNext (void) = 0;
-    virtual bool resetCandidate (unsigned int index) = 0;
+    virtual bool resetCandidate (size_t index) = 0;
     virtual bool unselectCandidates () = 0;
 
     virtual bool removeCharBefore (void) = 0;
@@ -106,8 +106,8 @@ public:
     virtual std::string restText (void) const = 0;
     virtual std::string auxiliaryText (void) const = 0;
     virtual std::vector<Candidate> candidates () const = 0;
-    virtual unsigned int cursor () const = 0;
-    virtual unsigned int focusedCandidate () const = 0;
+    virtual size_t cursor () const = 0;
+    virtual size_t focusedCandidate () const = 0;
 };
 
 }; // namespace PyZy
